src/backend/pyproject.toml
[tool.poetry]
name = "Poetryslam"
version = "0.1.0"
description = ""
authors = ["bharathpadmaraju <bharath.k.padmaraju@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
fastapi = "^0.60.1"
uvicorn = "^0.11.6"
starlette = "^0.13.6"
sqlalchemy = "^1.3.18"
sqlalchemy_utils = "^0.36.8"
python-multipart = "^0.0.5"
stringcase = "^1.2.0"
python-jose = {extras = ["cryptography"], version = "^3.2.0"}
passlib = {extras = ["bcrypt"], version = "^1.7.2"}

[tool.poetry.dev-dependencies]
pytest = "^5.4.3"
pycodestyle = "^2.6.0"

[tool.poetry.scripts]
server = 'PoetrySlam.server:server'

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
